mcp/requirements.txt
langchain
langchain-core
langchain-community
mcp
python-pptx
requests
duckduckgo-search
python-dotenv
groq